A rest interface to get the TaskIDs of a procedure is implemented using the Jersey Framework.
JSON de/serialization is done using the fasterxml Jackson Framework.

The entry point to the interface is rest/
To get the taskIds of the tasks assigned to a procedure the URL is:

rest/tasks/procedure/{procedureId}

The procedure should not have "/" in their id. If It has, the procedureId should be URLencoded and
tomcat should be configured with org.apache.tomcat.util.buf.UDecoder.ALLOW_ENCODED_SLASH = true (security concerns?).